Rewrite to ExpressionTypeResolverExtension
